+++ .github/release-notes/release-notes.md
@@ -3,6 +3,9 @@ Provides easier ways to generate randomized data.
 ## 🛠 Enhancements
 
 - generate recent dates based on a timespan.
+- add `Random<T>(int count)` to generate multiple data objects.
+- add `InRange<T>(int count, T min, T max)` and `InRange<T>(int count, IEnumerable<T> items)` to generate multiple data
+  objects.
 
 ## 📦 NuGet